Fix calendar red dot
Fixes #2 where a red dot would appear on a lend where it should have been green in the next month before today's date. This solves that issue by saying that **both conditions (month and date) need to be true ** in order to show the red dot, otherwise, it must be green.